refactor: move icons to its own folder
