+++ src/components/icons/IconChevronRight.vue
