+++ src/components/icons/IconConnpass.vue
