+++ src/components/icons/IconNote.vue
